Use a template to generate the 'whereto' testing data
As we add more branches and tags we don't want to manually add tests for
them (and we'd need to or the docs build will fail with an 'Untested
Rule'.

Use the redirections data that we use to generate the htaccess file to
also generate the tests.

Also I removed the debug output as it it's assertion is incorrect we can
actually get the .htaccess file from the docs job \o/

Change-Id: Ib5d35169e68cd92b666a35705d8e36942bd28b89