Release python-novaclient 13.0.0
This is a major version bump because of the 2.68 change:

https://docs.openstack.org/releasenotes/python-novaclient/unreleased.html#upgrade-notes

"Added support for microversion 2.68, which removes the --force option
from the nova evacuate, nova live-migration, nova host-evacuate and nova
host-evacuate-live commands."

Since the nova CLI negotiates the max microversion to use between the
client and server, if this is running against a Stein nova-api it will
remove the --force option from those commands, so someone whose script
is not pinned to a particular microversion and is using the
--force option would break. Granted, they can simply pass
--os-compute-api-version=2.67 to work around that, and probably should
have been pinning the requested version in scripts anyway to avoid
backward incompatible API changes on the CLI, i.e. the CLI behavior is
akin to passing microversion 2.latest which you shouldn't really do if
you're going for stability, and we have said in the past that the nova
CLI is not a bash SDK. So the major bump is possibly unnecessary, but
safe.

Co-Authored-By: Matt Riedemann <mriedem.os@gmail.com>

Change-Id: I1212ccc59dfb5d06f607d9f1f75a409bb8ef5500