Add cinderlib
This patch adds a new independent deliverable for openstack/cinderlib,
and will be changed to a cycle-based model once we change to Train.

The cinderlib project is a special kind of project, because even though
it's a library, it is unlike the rest of the OpenStack libraries.

It's primary use is for non OpenStack projects, and uses the Cinder
project itself as if it were a library to consume existing block storage
drivers.

Change-Id: I11c1637a1a09ed828117a480d9ff472ce9f5e277