Handle other rc cases
The new-release script couldnt handle cases where we wanted to
increment the version for an rc where there hadn't previously been
a release since milestones aren't as much of a thing anymore.

This patch will add the ability to detect and handle these cases
falling back to how things are done with milestones.

Change-Id: I7c7d6ad76c0b3f9aae41a45f2cd646e298966c6f