Move tarball-base settings to repo settings
This moves existing tarball-base settings out of per-release to the
overall repository-settings. This is nearly always the same for every
release, so it's a little redundant, and having it as an overall setting
will cause it to be carried over when running init-series and prevent
the need to remember to add it with new releases.

Change-Id: I6f5766313a5484653b881cf75f8b3c76a473f3e9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>