oslo.db 4.45.0
This is a necessary feature release off the stein branch because our
previous lower-constraint on psycopg2 does not work with postgresql 10,
which is now used in Ubuntu Bionic. The only change in this release
is to bump the minimum version of psycopg2 to one that does work
with postgresql 10.

Change-Id: I8d63fd0de8159b561314753d01e37384d73d58f3
Related-Bug: 1821400