Extract announce info from deliverable file
There were some undocumented expectations for formatting information in
repo's README files to be extracted by the email announcement scripts
that many (most?) repos are not following. Since we have information
like the bug reporting details in the deliverable files, this attempts
to parse the info from there.

Story: #2005279
Task: #30132

Change-Id: I31643d9c4f22c199c9fc385711fdf530c2836631
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>