placement: 1.0.0.0rc2 from stable/stein
Prepare an rc2 release of placement as a 1.0.0.0rc2. Three of the
four changes are administrivia. 0c9b38a8 is a bug fix in the
PlacementFixture which is used by nova (in stein and in master).
The bug was found in the window of time after placement had cut
stable/stein and before nova did.

No further backports are pending.

$ git log --oneline --no-merges 1.0.0.0rc1..bbd90ca2a671b76cda6844a3a8de5c46f587b7a5
bbd90ca2 (origin/stable/stein, gerrit/stable/stein) Replace openstack.org git:// URLs with https://
0c9b38a8 Add register_opts param to PlacementFixture
60efe8ac Update UPPER_CONSTRAINTS_FILE for stable/stein
de7c123c Update .gitreview for stable/stein

Change-Id: I9719f9a139755ac2848fb259c77e881350006e8a