Add TOC for release deliverables
With a large number of deliverables and different deliverable types, the
documentation output for a release series can get very long and it's not
always obvious that there are different groups of types if you scroll
all the way down.

This updates our deliverable extension to include a table of contents at
the top of the output to add convenient links to the different groups of
deliverables.

Change-Id: I3c55057320661f7167d44138941a07fd38c3c81f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>