Release 0.12.0 of os-traits
os-traits is a library of trait symbols used by the placement
service and clients of that service to represent qualitative
aspects of resource providers.

This change releases a 0.12.0 of the package, under the independent
deliverable style, switching from cycle related, because that's
what the team prefers. Being independent is considered low risk
because the library is made up of a list of strings which can only
be extended.

It also updates the team to be 'placement' and uses storyboard
instead of launchpad.

Change-Id: I84931cebe62d413f77ba0159a415db0ec04a2e7f