[horizon] final releases for pike
This is a procedural review to tag a final release for each deliverables

It's possible that the release isn't needed (for example if there are no code
changes since the previous release, sadly that's non-trivial to detect en masse.

If you are the PTL or release liaison for this team please +1 this review
if it is ready or -1 and take over the review to correct mistakes.

Don't tag DOA as there hasn't been any functional code change since
3.6.1

Change-Id: Ic460372fd8426b998d69690f4f5e737726c9f91d