[ironic] final releases for pike
This is a procedural review to tag a final release for each deliverables ; and

It's possible that the release isn't needed (for example if there are no code
changes sine the previous release, sadly that's non-trivial to detect en masse.

If you are the PTL or release liaison for this team please +1 this review
if it is ready or -1 and take over the review to correct mistakes.

Don't tag the following as there isn't any functional code chnages since
the last release:
 * ironic-inspector
 * networking-baremetal
 * sushy

Change-Id: I78721d822e5d113a4ebfea8ec59d8cc1a939043c