[tacker] final releases for pike
This is a procedural review to tag a final release for each deliverables ; and

It's possible that the release isn't needed (for example if there are no code
changes since the previous release, sadly that's non-trivial to detect en masse.

If you are the PTL or release liaison for this team please +1 this review
if it is ready or -1 and take over the review to correct mistakes.

Don't tag python-tackerclient or tacker-horizon as they don't contain
any functional changes from already tagged and released versions.

Change-Id: Ife5dbd016149b3585ff0ecc41868d2564901668e