[OpenStackSDK] final releases for pike
This is a procedural review to tag a final release for each deliverables ; and

It's possible that the release isn't needed (for example if there are no code
changes sine the previous release, sadly that's non-trivial to detect en masse.

If you are the PTL or release liaison for this team please +1 this review
if it is ready or -1 and take over the review to correct mistakes.

Don't create a new release for requestsexceptions as there aren't any
functional changes since 1.3.0

Change-Id: I9df31bd92ccf7613f708aef280024b8faf451bab