[OpenStackClient] final releases for pike
This is a procedural review to tag a final release for each deliverables ; and

It's possible that the release isn't needed (for example if there are no code
changes sine the previous release, sadly that's non-trivial to detect en masse.

If you are the PTL or release liaison for this team please +1 this review
if it is ready or -1 and take over the review to correct mistakes.

Change-Id: Id510eb5767775defdb903c5f7ecc887361699a04