[Telemetry] create pike-em tag against final releases
This is a procedural review to transition pike into extended
maintenance by creating the pike-em tag

It's possible that the release isn't needed (for example if there
are no code changes sine the previous release, sadly that's
non-trivial to detect en masse).

If you are the PTL or release liaison for this team please +1 this
review if it is ready or -1 and take over the review to correct
mistakes.

Change-Id: I8b58a6feac1a2d17a3ab6ac1d3daf7e6f25c1c3c