Fix release-note parsing of deliverable files
The list-changes job would fail to parse deliverable files with the
error "WARNING: Unable to parse $repo $series deliverable file" due to
the repo name being passed in with the leading openstack/ namespace.

This is common code used by list-changes and release-notes. In
list-changes we have the actual name of the deliverable file, so just
use that if it is available.

Also makes sure that cases were we expect just the repo name without
namespace only get what is expected.

Change-Id: Ib7526fc66d05530c2c2c722284c1d64a5242e419
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>