Move cinderlib to cycle-trailing
Cinderlib was added late in the Stein cycle, so it went in with the
independent release model, but the appropriate model is cycle-trailing.

Cinderlib is different from other OpenStack libraries.  It provides an
object oriented abstraction of the Cinder drivers for the purpose of
consuming them outside of OpenStack.  Using the Cinder drivers means
that it has the Cinder project as a dependency and needs it to be
released before the cinderlib release. This is why the cycle-trailing
release model is the most appropriate.

This patch adds cinderlib to the Train release as a cycle-trailing
project.

Change-Id: I4bc7068e2ac41af2b879d382e73cfd16e6f2927c