Release 0.4.0 of os-resource-classes
The placement team decided that os-resource-classes should
follow an independent release cycle as it eases various
aspects of integration with the multiple projects that
consume the package and the fact that the package is simply
a big enumeration of symbols pointing to strings.

The expectation is that henceforth there will be a new release
each time there is one or more new symbols available.

Change-Id: I6349fe9eaa65fe871f3ed13c8b066115c9a035c3