Add --interactive to new-releases
Shamelessly steal^Wcopy some of the code from interactive-release so
that after validating the release we list the changes that will be
released.  This gives us the ability to decide that the release
contains no functional changes and elect to not create that release.

NOTE: I chose not to use interactive-release as it seems not to
correctly handle first releases in a series (because it doesn't load all
release history.  It also doesn't use some of the new features (like
series_status).  Adding --interactive to new-release gets us a long way
to deprecating interactive-release but we aren't quire there yet.

Change-Id: I25bbb4d7df9ae618500dd37f4b0cbc32c0bbd153