Introduce cycle-automatic release model
For things like tempest-plugins, the cycle-with-intermediary model
is a bit overkill, especially now that we encourage multiple
intermediary releases to qualify.

Introduce a new 'cycle-automatic' model for deliverables that only
need to be released once, automatically at the end of the cycle.
This is limited to "other" or "tempest-plugin" deliverables.

Task: 31025
Change-Id: I83ff63cef18ae297013c3761a373078e580cf58b