Fix regression in list-changes
Recent merging of the liaison-loading code in list-changes makes
most list-changes job fail due to error in YAML loading code.

Beyond that YAML loading issue, the code assumes getting a dictionary
with team names as keys and an array of liaisons dictionary as values.
Fix the release_liaisons.yaml and schema to match.

Change-Id: I94fa41862d37aaf0535d64afd00b2bd55bb4649a