Skip branch checking on EOL deliverables
If we make modifications to a deliverable file for whatever reason and
that series is EOL, the validation code would still check that branches
existed and therefore fail. The stable/$series branch most likely is
deleted after going EOL, so the validation code would think there was a
need for a new branch to be created, then fail if the last release done
was not the same point as the branch.

This updates the validation logic to check up front if the deliverable
is EOL and skip these checks.

Change-Id: Ib5be3c0ba7283021301d278fcb362432f9a7d5cb
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>