Remove the stable/stein branch for Patrole
Patrole is branchless with "cycle-with-intermediary" model
and does not create the stable branch with same reason as
Tempest.

Below commit added the stable/stein for Patrole which is not
correct- https://review.opendev.org/#/c/650173

This commit removes the branching from stein patrole release.

As next step, we will add some type(as proposed by ttx[1]) or
tag to make it clear that Patrole does not need branch.

[1] https://review.opendev.org/#/c/670808/

Change-Id: Ib11838bc87e4fe0a350d107184be68e837233891