osc-placement 1.6.0
This library was not released since March. Libraries should be
regularly released to make their new features available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I0ee032f305c2f8969077f995bcff4e1610346448