python-aodhclient 1.3.0
This library was not released since March. Libraries should be
regularly released to make their new features available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I10c34bab1541b22818cc3294d9f13d2117665620