python-freezerclient 2.2.0
Master branch for this library was not released since December.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I57b24c372e6fe6fe7f2023225f9f08766f73740a