python-heatclient 1.18.0
Master branch for this library was not released since January.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: Icc057b1b9e02232fa45f012da7c6b6aee73cf1a5