python-karborclient 1.3.0
Master branch for this library was not released since January.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: Ie090a4d079f22b8536a21c3a90c708a8c3034ba8