python-pankoclient 0.7.0
Master branch for this library was not released since March.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I267ee9471630697358d1eec5219f621513c4c767