python-senlinclient 1.11.0
Master branch for this library was not released since March.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this

Change-Id: If660f12a1b6674a25f9c922f4f5939b636d63d7c