python-watcherclient 2.3.0
Master branch for this library was not released since October.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: Ia0b11e71005f020280681f3d40e9e4b149fa1fbd