shade 1.32.0
Master branch for this library was not released since March.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I7db9e25bd6953bed6cfb760f1d1f819d2db8be08