tosca-parser 1.5.0
Master branch for this library was not released since December.
Libraries should be regularly released to make their new features
available to consumers.

This change was generated semi-automatically. If there is a good reason
to hold on releasing this (like if you are waiting for a specific
feature or bugfix to land before releasing), please -1 this.

Change-Id: I4757970bba1f5274a870be90fefec8c57aa6ca95