Change blazar-dashboard release model
Since blazar-dashboard was only released once last cycle AND it has not
yet been released this cycle yet either, we propose to change it to the
cycle-with-rc model which is better for deliverables that do a single
release per cycle. 

There are three options for responding: 

1) If you know you will only do one release this cycle, +1.
2) If you know you will do more than one, please release now and -1 
this change.
3) If you have no idea when you will release, please please please
propose one BEFORE rc1 and -1 this change.


Change-Id: I0de7585358b636d4ad8b08d6f0063c5128ebe88d