Change heat-dashboard release model
Since heat-dashboard was only released once last cycle AND it has not
yet been released this cycle yet either, we propose to change it to the
cycle-with-rc model which is better for deliverables that do a single
release per cycle.

There are three options for responding:

1) If you know you will only do one release this cycle, +1.
2) If you know you will do more than one, please release now and -1
this change.
3) If you have no idea when you will release, please please please
propose one BEFORE rc1 and -1 this change.

Change-Id: Id970cbce3f716e221690d9c9522b6752b52118c1