Change tricircle release model
Since tricircle was only released once last cycle (and the cycle
before) AND it has not yet been released this cycle yet either,
we propose to change it to the cycle-with-rc model which is better
for deliverables that do a single release per cycle.

There are three options for responding:

1) If you know you will only do one release this cycle, +1.
2) If you know you will do more than one, please release now and -1
this change.
3) If you have no idea when you will release, please please please
propose one BEFORE rc1 and -1 this change.

Change-Id: I8b822f2b9c173fd72bc6015f69a3703a963f1c4c