Release openstack-doc-tools 2.0.0
This repo is only used for openstack-manuals and friends.

One of the changes publishes content to a different location, so we increase the major
version. This allows to simplify our CI scripts.

Also, removes some previously used tools that are now obsolete.

Increase major number since both are incompatible changes.

Change-Id: Iab9d4459ecd668c70ecb2a30fe1c9fecd9148516