Split OSA Queens into two files for consistency
OSA Rocky and above splitted OSA and OSA roles into two
deliverable files. This was to accomodate the fact that, during
Rocky, we decided to make the roles tagless, only branched.

However, splitting the files into two has another advantage:
We could now release separately the roles and the main repository.

This is still useful to do for Queens, as we regularly create
queens releases, even if we don't tag them anymore (because
we decided to not apply tags for roles into ALL branches).

By splitting in two files, requesting a new-release with the
openstack-ansible deliverable will do just the minimum. We would
still be allowed to "tag" the roles for EM.

Change-Id: I14f4894b1676377d40b10e77429ddeef68d360ff