Oslo releases for 2019-09-03
As we are now past Oslo feature freeze, this releases all projects
that have pending changes, including test and doc changes that
normally wouldn't require a release. This is so that when we branch
the projects we do so from a commit that includes all of the changes
this cycle.

Change-Id: Iae495750fe9d18b3dc8bce3e34e8ccd57851fe83