Release openstackclient 4.0.0
Note that this is the metapackage that installs python-openstackclient
and all of the plugins listed in requirements.txt at one time for user
convenience.

This is the first real release of this package, we are using major version 4
to correspond to the major version of python-openstackclient version that
will be installed.  We want to maintain that consistency for the major
version only and anticipate that the only other reason for releases will
be for adjusting the plugin list and their version requirements and do
not intend to match more than the major version with the primary OSC package.

Change-Id: Ic9c1e189385ec58fe834897f46dfbffbc00b1bb7
Signed-off-by: Dean Troyer <dtroyer@gmail.com>