Clarify release-test purpose
This is not clear what release-test is for.

I expected the repo to be used for final release, as a canary
test. I expected things had to done with it beforehand.
But I only got a confirmation through IRC (as this is not yet documented
in the process), that it only serves during RC/final release time.

This should clarify it.

Change-Id: I9ba636a8a3147466a1e661c074a050ff14a6d394