Sync ironic-prometheus-exporter releases
Ironic-prometheus-exporter 1.0.0 was recently released,
but there was already a 1.0.0 uploaded on Pypi, from June 13:

https://pypi.org/project/ironic-prometheus-exporter/1.0.0/

This is obviously very confusing: the recently-pushed 1.0.0
tag does not correspond to the published PyPI release. It also
prevented the release automation from publishing a 1.0.0 tarball
on our end.

Releasing 1.1.0 on the same SHA should sync things up.
Let's see if we can omit mentioning 1.0.0 in our release history too.

Change-Id: I58206822b06f0aaac4deb0e5b8ed99ba9692f075