Add stable-branch-type as 'none' for Tempest an Patrole
Tempest and Patrole are with cycle_with_intermediary model
but does not need stable branch due to their branchless nature.

There are many confusion over creating the stable branch for Tempest
and Patrole due to their current release model. To avoid that
and make it explicitly, new stable-branch-type 'none' has been
introduced.
- https://review.opendev.org/#/c/670808

This commit adding the stable-branch-type as none for Tempest
and Patrole.

Change-Id: I98b6862de741bbaeee20035d19b12783982d376c