Make Ussuri the future series
If we make Ussuri the future release, the htaccess will be generated
with redirections for it. When we decide that Train is maintained and
not developed, we should move Ussuri to development. And add 'Vancouver'
or something when we know the next series name to immediately generate
the new redirection.

Alternatively, we could change the code to generate the redirections
at all times. If the code is idempotent, we would never have an
issue of "forgotten" state of a series, and could bypass it.

Change-Id: I69008ebd330a90ae42166d55e5c06b694fc2d042