Send RC announces to release-announce
All releases are announced to the release-announce ML, except
Release candidates, which were still sent to openstack-discuss.

The rationale was that RCs should be announced to a developers
list rather than a downstream consumers list, so that they trigger
testing. But that is less true now that there is a single -discuss
list, where they generate a lot of noise around RC1, without
triggering any additional testing. They also confuse some downstream
consumers which expect those to go to the usual release announce list.

This patch removes the exception and makes sure we send all RC
announces to the release-announce list.

Change-Id: Id33dba37b4d53962a2170ec401499fe3dd2e24bf