Release sahara and sahara-plugin-* RC2 for Train
This release is being proposed to pick up translations and bugfixes
that have merged to stable/train but not been released.

Not all changes are critical to make it into the initial Train
coordinated release and can wait until a stable release later. If
this patch does not get an ack from the PTL and/or release liaison
by EOD Thursday it will just be abandoned.

Teams may also choose to -1 this patch and update it once any final
commits merge prior to the October 10 deadline.

Change-Id: I572379e09dd980abf475b3191cd0ce415cc6668d