Train final releases for cycle-with-rc projects
These are the final release taggings for all cycle-with-rc projects.
While not required, having PTLs ack this patch would be nice to include
that record in the merged metadata.

Since it always comes up - yes, the diff-start is right. That's the
calculated starting point where the last stable branch was created.

Change-Id: Icd4d3e6bc2bdd23b7073c38ea1a2e9100e679834
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>