Revert "Openstacksdk Queens 0.11.4"
The .gitreview file in stable/queens was not updated with
the new repo name, resulting in the tag job failing for
this release. Since it was not able to do anything, the
easiest fix is to just revert this patch, merge an update
to the .gitreview file [0], and resubmit it with the hash
including that change.

[0] https://review.opendev.org/689131

This reverts commit 9713aa37805c2d308a2a26c5f994b5c3a1adc854.

Change-Id: I75ebdb23613020624afeb33520e8c6d6eb15d186