Add note to highlights page explaining content
Some are linking directly to the highlights pages for release cycles as
a sort of top level release note page. This content is very raw and the
quality of the items listed really varied team by team.

This adds a note at the top of the pages to make sure it is clear to
anyone reading it that the content on the page is not something that can
be taken directly as the marketing information for a release.

Change-Id: I9bc3a94e0cd63735b0234f3cfb10817ee3b84c66
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>