Allow upper-constraints redirect based on SHA
This patch aims to make possible to provide SHA commit
for https://releases.openstack.org/constraints/upper/ URL
in addition to branch or tag. This is heavily used by OSA to be sure
that every time role is runned under some release it will install
the same package versions.
And to be consistent and use releases.openstack.org for contraints
not only in tox but in role itself this patch is needed.

Change-Id: I1ce0af2ab2163ea1f154516229889b3cf8813ab6