List the Ussuri cycle signing key as current
This sets the end date for the Train key, adds the Ussuri key
starting today and links to a minimal export of the corresponding
public key and its details on the keyserver network. This can be
amended with a new patchset or a follow-up change to correct the
included dates after the key material change (Depends-On below)
merges and goes into effect.

Depends-On: https://review.opendev.org/691939
Change-Id: Id8e4d24fa18cccffb2da3acd987cfbf01efd3bae