Release new neutron-tempest-plugin version
As we decided during Shanghai PTG for branch which is in EM we will
run neutron-tempest-plugin jobs with specific tag, not from master
always.
So lets release new version of neutron-tempest-plugin now and we will be
able to use this version for queens jobs instead of using still tests
from master branch.

Change-Id: Id8c7d2e6b52aee35ada9ee89708b08c12b8dc20c