Update process for R+1 week
Move R+1 week process items to the top of the document, remove obsolete
steps and merge the remaning ones with pre-PTG steps. Include weekly
email content. Move the step about proposing a schedule for the next
release to "between m-2 and m-3", as this is when we actually do that.

Change-Id: I5a4a7e5e059156139944a09e4d7813bd11ec1c5a