Change horizon release model to cycle-with-intermediary
It was discussed in the Shanghai PTG.
horizon has two aspects: the one is to provide GUI to users and
the other is a library for horizon plugins.
When horizon plugins would like to use recent changes or to avoid bugs,
they need to consume beta releases of horizon.
More frequent releases of horizon would make more sense.
This commit changes the release model to cycle-with-intemediary.

After this change, horizon major version will be bumped when both
user-visible features and horizon plugin interfaces are dropped.
This menas the horizon major version can potentially be bumped
multiple times in a single development cycle.

Change-Id: I494063ca132b362cb769468fc588d71ec6264ca6