Release horizon 17.0.0 (ussuri)
The parent patch proposes the release model change and
This is the first release for Ussuri.

Glance v1 support was dropped since the previous release
(i.e. Train release 16.0.0), so the major version is bumped.

This release would fix failures of the docs job in several horizon
plugins which is caused by the inconsistency between horizon 16.0.0
(train) and python-novaclient 16.0.0 (ussuri).

Change-Id: I72f79cfeea8e63812a18c8b09e31ed4df066d914