Do not list em and eol tags on releases site
We want these tags in git for developers and those working directly with
the source code, but we do not want to have these listed on the releases
site. Consumers of the site are just looking for actual official
releases they can use, so we should avoid confusion by not including
these "internal" tags.

Change-Id: I6a3fcc11efed6254ea6fb95579eafe5288508661
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>