Release heat-translator for Ussuri milestone 1
This picks up unreleased changes in heat-translator.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I061e44446373a787d4788e872da1cc6f195b00b9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>