Release metalsmith for Ussuri milestone 1
This picks up unreleased changes in metalsmith.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: Iefe3a49ed8a9bad34edc42d87e7dd1a0eba80afc
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>