Release mistral-lib for Ussuri milestone 1
This picks up unreleased changes in mistral-lib.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I25d8943227abbcf0b95624fd9c08d54c348e5041
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>