Release osc-lib for Ussuri milestone 1
This picks up unreleased changes in osc-lib.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I639aa13d9de4f0fd7b1023dc860934fb3429ad13
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>