Release oslo.db for Ussuri milestone 1
This picks up unreleased changes in oslo.db.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I52c555d2b967f67e14768dba5c5e5ae9e53bd5ad
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>