Release oslo.limit for Ussuri milestone 1
This picks up unreleased changes in oslo.limit.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I1afa11c61545ecc20e0df989841da6ed6c638f78
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>