Release oslo.reports for Ussuri milestone 1
This picks up unreleased changes in oslo.reports.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I52903f2f12fafcfe0ff7c6a72bf52142ae7290d4
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>