Release oslo.rootwrap for Ussuri milestone 1
This picks up unreleased changes in oslo.rootwrap.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I11dee309becb5902c624d0239a79c976abb071e1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>