Release oslo.service for Ussuri milestone 1
This picks up unreleased changes in oslo.service.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I679e0c544092d2de07658c6484c062449d6cb7f1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>