Release tosca-parser for Ussuri milestone 1
This picks up unreleased changes in tosca-parser.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: Iec878d5978840b981fe95fae826ee87e9f2c7a06
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>