Release ceilometermiddleware for Ussuri milestone 1
This picks up unreleased changes in ceilometermiddleware.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: Ib64198b265519397a51b3403aef384cf4816a898
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>