Release glance-store for Ussuri milestone 1
This picks up unreleased changes in glance-store.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: If8e94f9e39961c61d17110e8bffcd4057e07561a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>