Release ironic-lib for Ussuri milestone 1
This picks up unreleased changes in ironic-lib.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: Ia31c1d992e76fbf21b6360232efcd27b99c3eb66
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>