Release kuryr for Ussuri milestone 1
This picks up unreleased changes in kuryr.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: Ide89d1d975e94586f8f58c85baf97d53f0c713de
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>