Release monasca-common for Ussuri milestone 1
This picks up unreleased changes in monasca-common.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I8c00cd1fbfc1622647949991119652c119deee22
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>