Release monasca-statsd for Ussuri milestone 1
This picks up unreleased changes in monasca-statsd.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I2092e5b5b6d646c4b2e5dfc1ede65b7790cea39f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>