Release os-client-config for Ussuri milestone 1
This picks up unreleased changes in os-client-config.
The cycle-with-intermediary deliverables are good to release frequently
whenever there are functional changes to be released. Feel free to -1
this patch if there are additional changes that should be included in
this release or if it is not a good time to do it.

Change-Id: I342ac7abe7b43d1145b92d1c0da1911024b036b4
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>