governance/releases consistency check tool
Introduce a new tool to list inconsistencies between projects.yaml
in governance repository and deliverables in the releases repository.

Sometimes projects.yaml defines release-managed deliverables, but the
release team is not handling them. Sometimes, the release management
team handles deliverables which are not officially produced by an
OpenStack project team, and should probably not be listed on releases.o.o.

This tool lists both types of inconsistencies, and can be integrated
to release management process so that they are handled before becoming
an issue.

NB: Once ready, this tool could be used to replace the "membership
freeze check" tool which only looked at some of those inconsistencies.

Change-Id: Id263f0c91f872111fe997e2b8a1088b61033d114