Add signature links for independent deliverables
We have been filtering whether to include pgp links for deliverables if
the series was >= 'O', which resulted in filtering out any independent
deliverables.

This updates the check to account for independent deliverables. It also
makes the table format consistent for pre- and post-Ocata, just leaving
the column blank for series where we did not generate signatures.

Change-Id: I0c9f97cb12a429d2e54d94df2b3e110156278bc1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>