Ignore line length for bashate
I am not sure we really need to care about a line length that's 90
chars instead of 79. We have so many ugly things here, I am not
sure this will impact positively the readability to force line breaks.

Without this patch, all the shell scripts changes will fail with
a bashate failure, due to the fact that the
recently introduced changes in tools/make_missing_releases.sh
have lines of length above 79.

The alternative would be to fix those lines. There are only 3 of them.

Change-Id: I5fc7bde3c60c50ae78e3c7f297f715a7591e486c