Skip sig links if the skipped-sig flag is set
In releases.o.o we display signature links for all deliverables,
if the series is > Ocata or independent. Since some "independent"
deliverables predate the signature generation tooling, that results
in a number of "independent" deliverables displaying broken signature
links, which makes us look bad.

This adds a flag (skipped-sig) that can be set for independent
deliverables that did not have any signature generated (pre-Ocata), and
skips the signature link display if the flag is set.

As a practical example, this fixes broken links for PBR<2. Tony signed
up to automatically generate the others.

Change-Id: I44a49e3f08010a85c64673d2292528139eabcc99