Switch networking-hyperv to cycle-with-rc
Deliverables following the cycle-with-intermediary model are expected
to release multiple times throughout the development cycle. Since
networking-hyperv only released once last cycle, and has not done a release
yet this cycle, this proposes changing the release model to cycle-with-rc
so the expectation is only that it will release an RC by the end of
the cycle prior to doing the final release.

For the team, please respond to this patch in one of the following ways:

* -1 this patch and follow up with an intermediary release
* +1 this patch and we will switch over the release model
* -1 this patch, but acknowledge a release will be done prior to the
  RC1 deadline.

Change-Id: I16a6f7d2a1c3278cc46b09a1c7cebc2cc6d3e842
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>