[designate] Transition Rocky to EM
This transition the rocky branch to extended maintenance.
Changes for bugfixes and things the team deems important are
still encouraged, but there will no longer be official releases
off of the branch.

Please +1 if the team is ready for us to proceed with this
transition, or -1 if there are any final backports currently in
flight that we should wait for. For the latter case, please
update the patch with the new commit hash after doing a final
release to get those changes out so we know to proceed with the
transition.

Change-Id: If893e5c22d8c2f2f86ca6a472e74b457205ee2d8
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>