Don't fail version validation on install failure
There could be an issue in previous tagged releases due to external
factors. In that case, there is no way to fix that version.

Validation currently checks out the previous tag and creates an sdist in
order to compare requirements changes on bugfix version bumps only. In
the case where the previous release has an issue, just handle the
failure and move along.

Change-Id: I60db41a475c3a13359556198c0611489dffa4b3f
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>