Oslo releases for 2020-03-30
Major for castellan because one of the changes removes a public member
from some of their modules. Feature for oslo.vmware because I think
the change has an implied dependency bump on the Glance client, even
though the client is passed in and not explicitly listed in the
requirements file.

Change-Id: I9ad3b891c09dab47e38de4d77dcea6d3b1516bf8