Release final heat-translator for ussuri
Final ussuri non-client lib release for the April 2 deadline.

Please leave +1 if we are all set and can proceed. Leave -1 if we
need to wait a little bit longer for something final to land,
then update it with a better commit hash to use for the release
when ready.

Patches with no response will be approved a few days after the
deadline.

Change-Id: I6d79952aabef6bb5d21fd932eeca636fd83b48a1
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>