Release final mistral-extra for ussuri
Final ussuri non-client lib release for the April 2 deadline.

Please leave +1 if we are all set and can proceed. Leave -1 if we
need to wait a little bit longer for something final to land,
then update it with a better commit hash to use for the release
when ready.

Patches with no response will be approved a few days after the
deadline.

Change-Id: Ie486e6a25460b04a54f6f69e3afe74876bc31501
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>