Release final os-client-config for ussuri
Final ussuri non-client lib release for the April 2 deadline.

Please leave +1 if we are all set and can proceed. Leave -1 if we
need to wait a little bit longer for something final to land,
then update it with a better commit hash to use for the release
when ready.

Patches with no response will be approved a few days after the
deadline.

Change-Id: I3d4678bdb3b02a70e2c136c7f7fbc4513a085bff
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>