Release final ovsdbapp for ussuri
Final ussuri non-client lib release for the April 2 deadline.

Please leave +1 if we are all set and can proceed. Leave -1 if we
need to wait a little bit longer for something final to land,
then update it with a better commit hash to use for the release
when ready.

Patches with no response will be approved a few days after the
deadline.

Change-Id: I886aa583a5b38f498596f0b7fd3ac25a58ac75a5
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>