Final Oslo releases for Ussuri
All of these releases are for changes that would not normally require
a release. However, because the stable branches will be cut based
on the last release sha, we need to make sure there is a release
that includes them so things like docs and tests are not out of date.

Change-Id: Ica9d5824b33e3f244d4bff64d6f5d101f5b371fa