Release final python-qinlingclient for ussuri
Final ussuri client lib release for the April 9 deadline.

Please leave +1 if we are all set and can proceed. Leave -1 if we
need to wait a little bit longer for something final to land,
then update it with a better commit hash to use for the release
when ready.

Patches with no response will be approved a few days after the
deadline.

Change-Id: I35889d77ac855910274a96379f0e7f70cd2bab66
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>