Release bifrost for ussuri
This is the final release for bifrost for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: Ib8c393d542b1cb033f465fcfb02abcd7118da466
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>