Release cloudkitty for ussuri
This is the final release for cloudkitty for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I87bd5fbc41d51a14f738991bd446ef8e0a507549
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>