Release heat-agents for ussuri
This is the final release for heat-agents for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: Ie4f9b6f2786c02725c65214a4021a8b416aa0d00
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>