Release ironic-prometheus-exporter for ussuri
This is the final release for ironic-prometheus-exporter for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I7c2dc1d88466f5e29447eb8154e604536699fc50
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>