Release ironic-ui for ussuri
This is the final release for ironic-ui for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I1b8dd962fe6e51624223f73b04ad50d01a9e9b6e
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>