Release magnum-ui for ussuri
This is the final release for magnum-ui for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I3a399720ef90e291fbd04b0c8366b6ef2d576e76
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>