Release monasca-ui for ussuri
This is the final release for monasca-ui for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: Ie874eb5b081c905bd885b24a818adc4590e30aa7
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>