Release networking-baremetal for ussuri
This is the final release for networking-baremetal for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I9ff7764e78a1c16aac01bab7cbf44ad23018aa7b
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>