Release networking-generic-switch for ussuri
This is the final release for networking-generic-switch for the ussuri
cycle.

Please leave +1 if this is ready to go, or -1 if the team needs a
little more time. If so, update this path with the new commit
hash to use once ready.

Patches with no response will be processed soon after the April
23 deadline.

Change-Id: I55b7d9f5fe61d9795a387cd6a24d8dcb7d3e356c
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>