Tag and branch missed neutron RC1s
Somehow networking-midonet, networking-sfc, and neutron-fwaas were
missed with last week's RC1 deadline. This adds the RC1 tag and creates
the stable/ussuri branch for these repos.

Since this is already past the deadline, a quick +1 to confirm from the
team would be appreciated. Otherwise, this will be approved by April 30.

Change-Id: Ie0a7f3aedd642818aa60276feebb85ba5c4ead05
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>