Ussuri final releases for cycle-with-rc projects
These are the final release taggings for all cycle-with-rc projects.
While not required, having PTLs ack this patch would be nice to include
that record in the merged metadata.

Release activity should now be frozen until this is processed on May 13.
Any bugfixes merged to stable/ussuri can be queued up and ready to go as
a follow on stable release after the coordinated release date.

Since it always comes up - yes, the diff-start is right. That's the
calculated starting point where the last stable branch was created.

Change-Id: I66602ca7a6ab1af35bf33d4d82fe34191f2e3865
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>