Increase validate job timeout to 3 hours
This can take a long time now with doing the final release tag
validation when all deliverables are included in one patch. Part of the
delay appears to be that pbr was updated to build release notes if reno
is present. We may look for other ways to optimize the validation jobs
to speed things up, but for now we can just make sure we have a very
large window.

Change-Id: I248968bca9a90a7b06d4d6dd0644c50d926e5f45
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>