Speed up sdist build
We have a couple validations that need to build an sdist now. But once
it has been built in our temporary pristine clone, we don't need to
build it again. Because we create a clean clone, we also don't need to
worry about any stale dist builds. This adds a check to the sdist build
logic to recognize if an sdist has already been built and skip it.

There are also a few extra steps in the build process that are not
needed for our purposes. The biggest is the new(ish) pbr behavior of
building all release notes in the git tree if reno is present. This adds
an environment variable flag to tell pbr not to do that. While we're at
it, we don't need AUTHORS or CHANGELOG created either, so set the flags
telling pbr to skip those as well.

Change-Id: I15be9a12ff6684729980c40afae9625a8b8e550a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>