Release oslo projects for victoria 2020-05-29
These projects have new features or significantes fixes or changes.

These projects are not still released for victoria so we need to
release a minor version even if no features are included in changes.

Concerning oslo.cache the main changes are the functional tests that
could be useful for downstream maintainers.

Change-Id: Id7d99b76d4f6a95bb879c317481478f9bbbfc80a