Set ocata unmaintained transition date
Discussions happening about transitioning ocata to EOL. According to our
stable policy, the branch should go through an unmaintained period to
give others an opportunity to step up and maintain things. To make this
explicit, this sets a date to make the overall state of ocata
Unmaintained so there are no surprises if and when we decide to make it
EOL community-wide.

Change-Id: I90fd9fc31464a451cc5476578081061947df9c7a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>