Postpone Stein Extended Maintenance transition by 4 weeks
As it happened in the past, the Extended Maintenance deadline conflicts
with the Final Release of the actual development cycle. Same way as it
was resolved in that time, postpone the EM deadline by approximately 4
weeks, to a less busy period. Also move the deadline some days to not
to fall to weekend.

Change-Id: I2b502f6b5189bd0f7e6bf0286105ad27da1643c7