Handle degoverned teams in highlights extension
The highlights extension attempted to be helpful by pulling in team
details from governance into the generated highlights output. This did
not account for the possiblity that teams may actually leave official
OpenStack governance, and therefore may not have these details present
in the governance output.

This adds handling to fail gracefully in this scenario by just not
trying to include the information that is no longer there.

Change-Id: Iee97b9547a2044acd86864f788f7ba1294afbf0a
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>